+++ routes/tournaments.js
@@ -204,7 +204,7 @@ router.post('/admin/generate', function (req, res, next) {
     const body = req.body;
 
     const players = [];
-    ["group1", "group2", "group3", "group4"].forEach(key => {
+    ["group1", "group2", "group3", "group4", "group5", "group6", "group7", "group8"].forEach(key => {
         if (body[key]) {
             body[key].players.forEach(player => {
                 players.push({
@@ -215,7 +215,7 @@ router.post('/admin/generate', function (req, res, next) {
         }
     });
     const venues = {};
-    ["group1", "group2", "group3", "group4"].forEach(key => {
+    ["group1", "group2", "group3", "group4", "group5", "group6", "group7", "group8"].forEach(key => {
         if (body[key]) {
             venues[body[key].group.id] = body[key].venueId;
         }
@@ -313,13 +313,14 @@ router.post('/:id/player', function (req, res, next) {
 router.get('/:id', function (req, res, next) {
     axios.all([
         axios.get(`${req.app.locals.kcapp.api}/player`),
+        axios.get(`${req.app.locals.kcapp.api}/venue`),
         axios.get(`${req.app.locals.kcapp.api}/tournament/${req.params.id}`),
         axios.get(`${req.app.locals.kcapp.api}/tournament/${req.params.id}/overview`),
         axios.get(`${req.app.locals.kcapp.api}/tournament/${req.params.id}/matches`),
         axios.get(`${req.app.locals.kcapp.api}/tournament/${req.params.id}/statistics`),
         axios.get(`${req.app.locals.kcapp.api}/tournament/${req.params.id}/metadata`),
         axios.get(`${req.app.locals.kcapp.api}/match/modes`)
-    ]).then(axios.spread((playersResponse, tournamentResponse, overviewData, matchesData, statisticsResponse, metadataResponse, modesResponse) => {
+    ]).then(axios.spread((playersResponse, venueResponse, tournamentResponse, overviewData, matchesData, statisticsResponse, metadataResponse, modesResponse) => {
         const statistics = statisticsResponse.data;
         if (!_.isEmpty(statistics)) {
             statistics.checkout_highest = _.sortBy(statistics.checkout_highest, (stats) => -stats.value);
@@ -357,6 +358,7 @@ router.get('/:id', function (req, res, next) {
                         playoffs_matches: playoffsMatches,
                         statistics: statistics,
                         modes: modesResponse.data,
+                        venues: venueResponse.data,
                     });
                 });
             })).catch(error => {
@@ -373,6 +375,7 @@ router.get('/:id', function (req, res, next) {
                     matches: matches,
                     statistics: statistics,
                     modes: modesResponse.data,
+                    venues: venueResponse.data,
                 });
             });
         }