+++ src/components/matches-table/components/match-row/components/match-button.component.js
@@ -1,5 +1,5 @@
 module.exports = {
     onModalPressed(event) {
-        this.emit('show-modal', this.input.data);
+        this.emit('show-modal', this.input.data, this.input.modal);
     }
 }